requirements.txt
ips.py==0.1.2
Pillow==9.5.0
